Figured out why CI was failing intermittently.
Since ssh_interface was unset, the default behavior tries to use
the public ip but falls back to the private ip if it doesn't see a
public ip. It is most likely the case that describe-instance was
not returning a public ip on the first attempt, so it was falling
back to the private. Forcing it to public should make it try again
until it has a public ip.