Fixed a bug uncovered in testing.
It turns out that sourcing sdkman's init script is necessary
because systemd does not run the command in an interactive
environment and therefore does not call .bashrc. Also,
hugepages.sh does not appear to be working according to test
results despite returning 0, so it was altered to check its own
work and exit accordingly.